Xorshift Random Number Generators

Xorshift random number generators, also called shift-register generators, are a class of pseudorandom number generators that were invented by George Marsaglia.They are a subset of linear-feedback shift registers (LFSRs) which allow a particularly efficient implementation in software without the excessive use of sparse polynomials. They generate the next number in their sequence by repeatedly taking the exclusive or of a number with a bit-shifted version of itself. This makes execution extremely efficient on modern computer architectures, but it does not benefit efficiency in a hardware implementation. Like all LFSRs, the parameters have to be chosen very carefully in order to achieve a long period.

xorshift_32(seed, n)

xorshift_64(seed, n)

xorshift_128(seed1, seed2, seed3, seed4, n)

Arguments

seed

Integer-like scalar used to initialize the internal state for xorshift_32() (32-bit variant) and xorshift_64() (64-bit variant). Different seeds produce different sequences. A seed of 0 is permitted for the 32- and 64-bit variants, but may yield a degenerate sequence depending on platform integer behavior.

n

Integer scalar. Number of pseudorandom values to generate. Must be non-negative.

seed1, seed2, seed3, seed4

Integer-like scalars used to initialize the four-word internal state for xorshift_128(). The state must not be all zeros; if all four seeds are 0, the function errors.

Value

A numeric vector of length n containing the generated values. Values are returned as type numeric (double) because that is R's default numeric type.

Details

For execution in software, xorshift generators are among the fastest PRNGs, requiring very small code and state. However, they do not pass every statistical test without further refinement. This weakness is amended by combining them with a non-linear function, as described in the original paper. Because plain xorshift generators (without a non-linear step) fail some statistical tests, they have been accused of being unreliable.

For more information, check out the Wikipedia Page, https://www.jstatsoft.org/article/view/v008i14 and https://stackoverflow.com/questions/71522815/why-does-the-xorshift-random-number-generator-always-seem-to-use-these-specific

These functions implement classic xorshift generators using bitwise XOR and shifts. They are intended for speed and reproducibility, not cryptographic security.

See also

Examples

xorshift_32(seed = 51966, n = 1000)
#>    [1]  1.373633e+10  3.567692e+15  2.821845e+18  3.963215e+18  9.079873e+18
#>    [6]  4.142066e+18  3.393284e+18  6.828410e+18  6.954072e+18  5.182148e+17
#>   [11] -8.839640e+18  7.017160e+18 -6.143611e+18  4.995252e+18  8.492982e+18
#>   [16] -1.361630e+18  7.539703e+17  2.249863e+18 -6.650698e+18  5.539515e+18
#>   [21] -1.729996e+18 -1.531593e+18 -6.058377e+18  1.981741e+18  4.738472e+18
#>   [26] -8.518671e+18  2.666515e+18 -2.663897e+18 -4.649928e+17  5.074701e+18
#>   [31] -3.666286e+18  5.433401e+18 -8.552317e+18 -2.037910e+18 -1.864380e+18
#>   [36] -4.741339e+18 -7.984653e+18 -1.418304e+18 -9.139444e+17 -6.307620e+18
#>   [41]  3.384515e+18  3.417787e+18 -6.808410e+18  5.704682e+18 -6.829719e+18
#>   [46] -6.761498e+18 -3.755897e+17  1.953470e+18 -9.156323e+18 -8.558197e+18
#>   [51]  9.196892e+18  3.784222e+17  1.376821e+18 -3.397189e+18 -7.351779e+18
#>   [56]  8.098128e+18  7.445951e+18  8.269929e+18  9.141497e+18  8.954985e+18
#>   [61] -4.093594e+18 -5.537952e+18 -8.654460e+18  7.365144e+18 -7.384471e+18
#>   [66]  6.334699e+18 -1.518345e+17 -7.049105e+18 -3.491732e+17 -8.996963e+18
#>   [71]  1.969053e+18 -7.319708e+17 -5.906078e+18 -2.986331e+18  7.287910e+18
#>   [76] -3.053878e+18  8.024131e+18  2.970456e+18 -7.905076e+18 -5.250758e+18
#>   [81]  6.798747e+18  6.152668e+18  3.979784e+17 -7.699275e+18  6.645202e+18
#>   [86]  3.137633e+18  5.433594e+18  2.720161e+18 -6.678385e+18 -7.646268e+18
#>   [91] -8.574604e+18  4.515430e+18 -5.291309e+18 -5.545759e+18 -1.822044e+18
#>   [96] -2.726144e+18 -1.331962e+18 -8.266801e+18  6.465513e+18  6.539394e+18
#>  [101]  8.170043e+18  3.312303e+18 -3.461937e+18  5.356181e+18  8.703330e+18
#>  [106] -5.840121e+18 -1.907917e+18  4.567581e+18 -8.754948e+18 -1.851856e+18
#>  [111]  4.065125e+18 -7.526038e+18  3.313196e+18 -7.913149e+18  7.942970e+18
#>  [116] -5.815805e+18  8.617516e+18  3.706793e+18 -3.795747e+18 -3.944074e+18
#>  [121] -3.582547e+18 -9.445005e+17 -5.912500e+17 -5.399478e+18 -2.080152e+17
#>  [126]  3.242088e+18 -5.913871e+18  8.200608e+18  2.458174e+18  8.869831e+18
#>  [131]  3.302430e+17  3.928425e+18  3.705076e+18  8.161697e+18  8.299692e+18
#>  [136]  7.889528e+18 -7.668535e+18 -3.860043e+18  3.625166e+18  3.266974e+18
#>  [141]  3.098278e+18 -5.697510e+18 -1.918776e+18  2.867073e+18  7.999191e+18
#>  [146] -8.947359e+18  6.455648e+17 -6.822738e+18 -9.020936e+18 -8.328273e+17
#>  [151] -8.302943e+18 -4.608907e+18 -8.692346e+18 -1.572933e+18 -7.818267e+18
#>  [156] -4.724981e+18  2.785587e+18  7.715915e+18  5.138170e+18 -6.088254e+18
#>  [161]  3.029913e+18  4.713485e+18 -1.749316e+17  1.006861e+18 -3.848460e+18
#>  [166]  8.155290e+18  6.632776e+18  4.423379e+18  3.560124e+18  8.612815e+18
#>  [171]  2.958060e+18  2.497935e+18 -9.136581e+17 -3.371332e+17 -3.038550e+18
#>  [176]  8.012058e+18 -6.205343e+18  1.678210e+18  8.381943e+18  4.509554e+18
#>  [181] -2.915012e+18  8.048098e+18 -3.433076e+18 -4.622043e+18 -9.728325e+17
#>  [186]  8.962350e+18  4.668213e+18  6.435157e+17  4.246531e+18  7.463681e+18
#>  [191]  6.199251e+18  4.205189e+18  6.880443e+18 -3.530763e+18 -1.020832e+17
#>  [196] -3.946474e+18 -4.882591e+18 -4.847451e+18 -9.660315e+17 -8.991690e+18
#>  [201]  7.189460e+18 -8.764633e+18 -1.847531e+18 -6.427562e+18  3.645340e+18
#>  [206] -2.343030e+18 -4.153232e+18 -7.868600e+18  2.259295e+18  2.657763e+18
#>  [211] -1.855549e+18 -2.960107e+18  2.451185e+18 -8.884622e+18 -7.793283e+18
#>  [216]  3.986576e+18  2.207333e+18 -7.203877e+18  8.593507e+18 -4.350616e+18
#>  [221]  5.224417e+18  3.681863e+18  2.950294e+18  3.803056e+18 -1.907764e+18
#>  [226] -7.625740e+18 -7.424621e+18 -9.082406e+18  1.792817e+18  6.613295e+18
#>  [231]  5.997649e+18 -1.037618e+18  8.741666e+18 -3.692668e+18  9.029216e+18
#>  [236] -3.632400e+18 -2.227362e+18  6.809100e+18  7.756577e+18  2.015182e+18
#>  [241] -2.035435e+17 -5.023142e+18  8.065501e+18 -7.114373e+17 -2.100810e+18
#>  [246]  4.730743e+17  6.465594e+17  2.291041e+18  3.277430e+18 -1.179477e+18
#>  [251] -5.701808e+18  2.811369e+18  5.578320e+18  1.713185e+18 -1.052250e+18
#>  [256] -3.968774e+18  3.175422e+17  4.881197e+17 -8.938696e+18 -1.454323e+18
#>  [261]  7.073039e+18 -4.122092e+18 -1.092684e+17  8.982897e+18  8.725630e+18
#>  [266]  3.058218e+17 -5.672045e+18 -1.794532e+18  4.345377e+18 -8.318398e+18
#>  [271] -8.590686e+18  7.685557e+18  8.749743e+18 -6.168833e+18  2.257978e+16
#>  [276] -1.302214e+18 -8.165067e+18 -3.897331e+18  5.765536e+18  5.954697e+17
#>  [281]  5.991025e+18  3.298688e+18  5.787054e+18  6.152423e+18 -1.022474e+18
#>  [286] -1.874872e+18  5.669169e+18 -4.970063e+18  3.091461e+18 -2.648047e+18
#>  [291]  5.343399e+18 -5.100941e+18  7.002188e+18 -3.974666e+18 -5.771418e+18
#>  [296] -7.883021e+18  8.805605e+18  7.595818e+18  7.971947e+18 -3.071883e+18
#>  [301]  6.598657e+18 -4.564129e+18  2.293864e+18 -8.128918e+18  4.619330e+18
#>  [306] -5.128924e+18  8.419106e+18 -8.362249e+18 -3.241728e+18 -7.642383e+18
#>  [311]  4.240905e+16  7.791348e+18  5.927554e+18 -2.721661e+18  3.028949e+18
#>  [316] -8.676137e+18 -2.470980e+18 -9.078746e+18 -4.994979e+18  3.806533e+18
#>  [321] -4.151392e+18  5.228306e+18  4.494069e+18 -4.618178e+18 -2.021805e+18
#>  [326] -3.180093e+18  4.652168e+18 -1.324337e+18 -5.463629e+18 -5.612839e+18
#>  [331]  6.100953e+18  6.825309e+18 -3.413701e+17  2.546495e+18  8.562379e+18
#>  [336]  7.997603e+18 -3.428984e+18 -3.328535e+18 -7.516688e+18 -4.847281e+17
#>  [341] -8.392340e+17 -7.426005e+17  8.748869e+18 -7.937958e+17  7.776201e+18
#>  [346]  3.479940e+18 -5.354733e+18 -5.944033e+18  8.853266e+18 -2.948631e+17
#>  [351]  3.844670e+18 -9.010591e+17  7.772483e+18  2.099820e+18  4.615991e+17
#>  [356] -9.173027e+18 -4.674072e+18  9.218188e+18  7.320989e+18  9.148666e+18
#>  [361]  3.221598e+18 -4.618059e+18  4.546264e+17 -8.378253e+18  7.796584e+18
#>  [366] -7.672515e+18 -7.083512e+18  6.785033e+18 -5.986126e+18 -7.567163e+18
#>  [371] -1.369230e+18 -7.852640e+18  1.024252e+18  7.921736e+18 -9.216680e+18
#>  [376]  6.938532e+18  5.495458e+18 -7.709442e+18  6.534287e+18  9.964803e+17
#>  [381] -1.746429e+18 -3.735787e+18 -8.409872e+17  2.487403e+18 -3.726725e+18
#>  [386]  5.682871e+18  5.973171e+18  5.492192e+18  8.322645e+18  4.062973e+17
#>  [391]  7.554998e+17 -8.370786e+18 -7.615123e+18 -6.035882e+16  1.442755e+18
#>  [396] -1.579026e+18 -2.361253e+18 -3.662733e+18  7.789714e+18 -4.373994e+18
#>  [401] -1.566236e+18 -6.494451e+18 -5.684616e+18 -6.229465e+18 -1.196185e+18
#>  [406] -1.877470e+18 -1.372572e+18 -7.738675e+18 -6.966593e+18 -9.153359e+18
#>  [411]  8.794602e+18  7.106823e+18  2.408477e+18  1.959717e+18  8.605411e+18
#>  [416]  3.919793e+18 -2.700867e+18  8.761042e+18  5.350483e+17  1.096013e+18
#>  [421]  1.961210e+18 -6.080969e+17 -7.400310e+18  8.065415e+18  2.307605e+18
#>  [426] -2.240566e+18 -6.554178e+18 -3.418708e+18  3.771852e+18 -8.257681e+18
#>  [431] -5.732347e+18 -7.203982e+18 -7.652766e+18 -6.029094e+17  7.386147e+18
#>  [436] -2.817493e+18  1.080505e+18 -2.216414e+18 -5.898987e+18  7.078757e+18
#>  [441] -8.260197e+18  1.629436e+18 -7.152694e+17  5.750582e+17 -8.527572e+18
#>  [446] -6.277856e+18 -1.319052e+18  5.162231e+18  8.683939e+18  3.090769e+18
#>  [451] -7.191747e+18  1.170790e+17  5.135919e+17  7.203118e+18 -4.963723e+18
#>  [456]  3.516369e+18  2.852010e+18 -2.786086e+18  2.018425e+18  5.444037e+18
#>  [461] -9.809290e+17 -7.015741e+17  5.060467e+18  4.780129e+18  8.287347e+18
#>  [466]  5.001005e+17  1.504008e+18  6.706235e+18 -9.094938e+18 -7.446011e+18
#>  [471]  1.274657e+18 -2.308890e+18  3.512354e+18  7.315787e+18 -7.129230e+18
#>  [476]  5.523496e+18 -5.851154e+18 -2.084325e+16  8.925125e+18  2.710774e+18
#>  [481]  3.206774e+18 -6.542228e+18 -9.168434e+17 -7.913517e+18  5.691235e+18
#>  [486] -6.266603e+18 -6.808667e+18 -3.506739e+18 -6.169605e+18  6.929404e+18
#>  [491] -6.964373e+18  8.186717e+17 -6.736701e+17  7.242816e+18  1.951477e+18
#>  [496] -1.674492e+16 -7.108954e+18  8.894705e+18  4.797813e+18 -4.668536e+18
#>  [501] -9.008241e+18  7.303212e+18  7.499561e+18  4.161822e+18  3.843914e+18
#>  [506] -2.830017e+18 -1.309039e+17  6.359102e+18  1.563785e+18  7.025456e+18
#>  [511]  2.288994e+18 -1.326837e+18  1.783195e+18  5.814906e+18 -2.915100e+18
#>  [516] -5.169405e+18 -5.344437e+18  7.213270e+18  6.579658e+17  6.809855e+18
#>  [521]  9.052362e+18  8.028464e+17 -4.273752e+18  1.506541e+18 -4.964259e+18
#>  [526]  1.507897e+18 -5.815743e+18  6.741136e+18  6.688638e+17  6.570351e+17
#>  [531] -1.927408e+18 -6.723675e+18  3.747780e+18 -3.871197e+18 -6.252724e+18
#>  [536] -4.039726e+18  3.622681e+18 -3.124159e+18 -3.352622e+18 -4.760854e+18
#>  [541] -1.073285e+18  4.311468e+18  5.274461e+18 -6.745056e+18 -8.822595e+18
#>  [546] -5.907117e+18  7.951491e+17 -4.000251e+18  5.438489e+18 -7.296726e+18
#>  [551]  2.225957e+18 -8.881422e+18 -6.915275e+18 -6.370530e+18 -1.427943e+18
#>  [556]  2.466640e+17 -6.136442e+18 -4.259261e+18 -7.028676e+18  8.456678e+18
#>  [561] -4.225955e+18 -5.745348e+18 -8.663093e+18 -6.436206e+18 -5.712414e+18
#>  [566]  4.385643e+18 -8.750849e+18  8.346310e+18 -5.041374e+18  1.589588e+18
#>  [571]  6.556750e+18 -3.771557e+18  6.651376e+18 -6.611606e+18 -7.766236e+18
#>  [576]  3.338374e+18  5.312020e+18  3.123008e+18  4.079765e+18  2.168251e+18
#>  [581] -8.498599e+18 -5.755337e+18  1.528099e+18 -5.195466e+18  7.301794e+17
#>  [586]  8.428310e+18  8.979021e+18  2.549295e+17 -9.172125e+18  1.602705e+18
#>  [591] -6.894486e+18 -2.002556e+18  3.837458e+18  2.449819e+18 -1.699158e+18
#>  [596]  8.254752e+17  5.006001e+18  8.891073e+18 -4.079693e+18  8.802021e+18
#>  [601] -2.026065e+18  4.709068e+18 -5.307232e+18 -3.184664e+18  2.812417e+18
#>  [606] -1.569560e+18  6.938477e+18 -8.400316e+18  3.940076e+18 -8.817186e+18
#>  [611]  1.359741e+18 -9.079202e+18  2.657384e+17  3.013930e+18 -2.012121e+18
#>  [616]  2.132805e+18  8.353313e+18  2.597982e+18 -7.815073e+18 -6.277875e+18
#>  [621] -5.937402e+18  3.274554e+18 -8.889806e+18  8.305911e+18 -7.870743e+18
#>  [626]  9.185474e+18 -3.112900e+18 -2.476340e+18  3.786827e+18  2.168728e+18
#>  [631] -4.783515e+18 -7.669477e+18  1.778289e+18 -5.410752e+18 -7.662627e+18
#>  [636] -5.065372e+18 -5.326962e+18  9.086683e+18  8.743307e+18  4.630802e+18
#>  [641] -7.923347e+18 -6.048936e+17 -6.143995e+18 -1.372042e+18 -2.418960e+18
#>  [646]  6.028167e+18  5.419978e+18  6.690486e+18  7.891111e+18 -1.658040e+18
#>  [651] -8.233951e+18 -2.817653e+18 -3.499025e+18 -2.197679e+18 -4.388431e+18
#>  [656]  2.407971e+18  4.231302e+17  6.696697e+18 -2.009766e+18 -7.363642e+18
#>  [661]  1.967156e+18  1.319033e+18  7.334285e+17  6.899684e+18 -2.542896e+18
#>  [666] -5.292714e+18 -7.415156e+17 -7.029990e+18 -8.462571e+18 -6.401142e+18
#>  [671]  6.673568e+18  9.020180e+18  7.850135e+18 -4.221861e+18 -4.023848e+18
#>  [676]  7.900293e+18  6.859591e+18  6.419117e+18 -6.863123e+17 -1.996397e+18
#>  [681]  8.841769e+18 -4.891308e+18 -5.065793e+18  1.780838e+18 -1.465845e+18
#>  [686] -9.197011e+18 -6.637684e+18  5.739074e+18  3.851880e+18 -8.251265e+18
#>  [691] -1.684181e+18 -5.438499e+18 -4.673754e+18 -4.401047e+18  5.690105e+18
#>  [696]  4.315069e+18 -3.198190e+18 -1.488535e+18 -3.336174e+18 -8.551601e+18
#>  [701] -9.088817e+18  6.411092e+18  5.079215e+18 -1.572221e+18  6.241954e+16
#>  [706] -6.065931e+18  1.210332e+18 -6.004047e+18 -3.968725e+18  6.380264e+18
#>  [711] -3.320396e+18 -3.998971e+18 -8.908048e+17  5.694980e+17  2.216261e+18
#>  [716]  4.787068e+17 -7.944876e+18 -1.868959e+18  6.356173e+18  5.485761e+18
#>  [721] -5.015683e+18 -4.960893e+18  5.502194e+18  5.819480e+18 -2.305868e+18
#>  [726]  8.674373e+18  4.378806e+17  1.835140e+18  2.963757e+18 -6.751073e+18
#>  [731]  8.865566e+18  4.967386e+18 -2.017899e+18 -5.957617e+18  7.465391e+18
#>  [736]  5.693765e+18  6.973448e+18  3.548156e+17  3.676740e+18  4.439037e+15
#>  [741] -1.422987e+18  5.855922e+18  5.949600e+18 -4.508082e+18 -6.900261e+18
#>  [746]  7.584504e+18  4.322348e+18 -8.717453e+18  4.612162e+18 -5.406744e+18
#>  [751]  5.558172e+18 -2.670875e+18 -6.709552e+17  1.055064e+18  8.988297e+17
#>  [756] -8.569340e+18 -8.941264e+18  5.048905e+18 -7.900370e+18 -6.756748e+18
#>  [761]  6.071073e+18 -1.981779e+18 -5.055834e+18  4.832757e+17 -8.316391e+18
#>  [766]  2.502992e+18  6.880129e+18  5.812137e+18  3.371676e+18 -4.533299e+18
#>  [771] -2.472326e+18  6.376714e+18  1.372848e+18 -8.200000e+18 -4.477862e+18
#>  [776]  2.886272e+18 -8.546334e+18  5.141131e+18  7.594284e+18 -8.364119e+18
#>  [781] -3.063829e+18  5.071348e+18 -7.332446e+18 -3.788245e+18 -4.235708e+18
#>  [786]  7.807181e+18  5.145482e+18  1.682847e+18  1.936739e+18 -1.442148e+18
#>  [791] -3.628827e+18  5.155267e+18  7.890780e+18  5.956584e+18  7.049053e+18
#>  [796]  2.516578e+18  6.020195e+18 -4.773041e+18  6.647313e+18 -2.683527e+18
#>  [801]  1.020331e+18  1.310145e+18 -4.425917e+18  1.252081e+17 -2.273495e+17
#>  [806]  5.044030e+18 -2.382482e+17  1.360869e+18  5.769033e+17  8.097614e+18
#>  [811]  7.598464e+18 -8.364242e+18 -8.623639e+18  8.007328e+18 -8.992238e+18
#>  [816] -2.468959e+18  1.827262e+18 -5.308352e+18  2.637785e+18 -1.127200e+18
#>  [821]  1.427284e+18  6.213781e+18 -3.889669e+18  5.925623e+18 -4.683001e+18
#>  [826] -8.857376e+18 -5.567963e+18 -3.663792e+18 -2.948368e+17 -8.640745e+18
#>  [831] -6.241277e+18 -3.956483e+18 -2.159045e+18 -5.612775e+18  3.887393e+18
#>  [836] -8.912392e+16  3.780216e+18  7.180073e+18  8.659342e+18  6.776283e+18
#>  [841]  2.053197e+18 -1.235760e+18 -3.876518e+18 -6.063053e+18  8.356915e+18
#>  [846] -3.201009e+18  4.240277e+18 -7.407742e+18  6.172301e+18  4.633875e+18
#>  [851] -4.066500e+18 -5.050165e+18  1.541642e+18  1.549965e+18 -6.689853e+18
#>  [856] -6.139013e+18  2.961031e+18  3.764611e+18  3.762416e+18  5.525444e+18
#>  [861]  6.530167e+18 -5.640652e+18  6.408306e+17  4.775425e+18  7.122598e+18
#>  [866]  1.909988e+18 -1.123431e+18 -1.071649e+18  4.034406e+18  3.933521e+18
#>  [871] -8.998180e+18 -3.657574e+18  6.959584e+18  9.128792e+18 -4.983300e+18
#>  [876]  1.364241e+18 -6.946648e+18  5.593244e+18 -2.455776e+18  6.458160e+18
#>  [881]  8.053782e+18  3.451673e+18  4.408567e+18 -3.068378e+18 -7.186316e+18
#>  [886] -4.006663e+17 -7.677893e+18 -2.240357e+17 -6.876436e+18  8.497270e+18
#>  [891]  4.245685e+18 -2.432237e+18  1.348268e+18 -8.317112e+18 -3.430311e+18
#>  [896]  2.841893e+18  9.016639e+18 -3.681097e+18  7.683499e+18  1.284189e+17
#>  [901] -3.602856e+18  6.583980e+18  1.512688e+18 -2.922484e+18  1.157540e+18
#>  [906] -4.771942e+18  8.198321e+18 -6.390528e+18  7.439164e+18 -4.110609e+18
#>  [911]  2.859399e+18  9.084823e+18 -2.457311e+18  6.882038e+18  3.964075e+18
#>  [916]  6.143920e+18 -3.140956e+18  2.946441e+18  4.488125e+18 -4.578997e+18
#>  [921] -4.593211e+18  8.133914e+18 -6.963321e+18 -7.081713e+18 -1.149466e+18
#>  [926]  7.349637e+18 -6.828609e+18  6.483787e+18  3.361839e+18 -7.729416e+18
#>  [931]  1.655428e+18 -1.516429e+18  7.353615e+18  2.780538e+18 -1.722174e+18
#>  [936] -5.295411e+18 -7.604844e+18  3.895364e+18  7.448054e+18  2.310297e+18
#>  [941] -7.589324e+18 -3.211278e+18 -6.351922e+18 -5.595970e+18  7.481606e+18
#>  [946]  1.818029e+18 -2.570350e+18 -1.575785e+17 -1.263472e+18  4.286734e+18
#>  [951] -4.456149e+18 -4.244971e+18 -1.245296e+18 -8.486558e+18  4.055001e+18
#>  [956] -9.774798e+16  6.308421e+18  2.943256e+18  1.099459e+18 -1.891830e+18
#>  [961] -9.935556e+17 -3.995535e+18  8.189419e+18  3.693579e+18 -7.007407e+17
#>  [966]  1.261845e+18 -7.351861e+18  8.990224e+18  6.304177e+18 -7.274120e+18
#>  [971] -2.886173e+18  7.330895e+18  5.505014e+17 -7.824076e+18 -6.556146e+18
#>  [976] -6.808499e+17  6.500484e+17  4.215318e+18 -3.837275e+18  8.981712e+18
#>  [981] -8.389387e+18  4.698658e+17  5.620651e+18 -2.532077e+18 -3.766020e+18
#>  [986] -7.904932e+18  8.335117e+18 -6.028628e+18 -1.404530e+18  5.809224e+18
#>  [991] -7.416506e+18 -3.928901e+18  2.680911e+18  1.926192e+18 -1.592285e+17
#>  [996]  3.799084e+18 -4.700297e+18  3.362514e+17 -6.254937e+18 -3.528217e+18
xorshift_64(seed = 51966, n = 1000)
#>    [1]  5.591807e+13 -3.261497e+18  3.209968e+18  6.832877e+18 -7.995511e+18
#>    [6]  4.644701e+18 -8.850609e+18 -3.194078e+18  9.860530e+16  9.101883e+18
#>   [11]  1.643804e+18 -3.872690e+18  2.115252e+18 -1.311833e+18  8.028923e+18
#>   [16] -6.716536e+18  2.979749e+18 -2.798426e+18  7.873278e+18 -7.152858e+18
#>   [21]  7.603477e+18 -2.687743e+17 -5.660034e+18 -1.393520e+18 -6.798714e+18
#>   [26]  1.235447e+18 -6.121882e+17 -6.293560e+18  7.987531e+18  1.661269e+18
#>   [31]  6.944849e+18 -5.871248e+18 -5.378124e+18 -5.376034e+18  2.794200e+18
#>   [36] -5.919191e+18  1.013009e+18 -8.597908e+18  2.049905e+18 -3.322626e+18
#>   [41] -2.190135e+18  5.403202e+18  3.719822e+18  4.657470e+17  7.355280e+18
#>   [46] -4.221077e+18  2.752891e+18 -5.753785e+18  8.287693e+17  6.655159e+18
#>   [51] -3.338360e+18 -4.921823e+18  5.635462e+17  1.321595e+18 -7.675861e+18
#>   [56]  3.803711e+18 -1.275919e+18  5.812251e+18  7.803638e+18 -5.274622e+18
#>   [61]  4.619259e+18  8.901453e+17  3.956382e+18 -4.537414e+17  6.761886e+18
#>   [66]  2.348732e+18 -8.871680e+18  4.182037e+18  1.123961e+18 -4.371580e+18
#>   [71]  7.748677e+18  6.398753e+18 -3.507537e+17  3.282331e+17  7.377398e+18
#>   [76]  6.029015e+18 -3.854433e+18 -5.236688e+18 -3.618284e+18  2.874339e+17
#>   [81] -1.584563e+18  3.393622e+18  5.053877e+18  8.915159e+18 -8.373599e+18
#>   [86] -4.030094e+16 -6.963107e+18  2.677232e+18  8.420222e+18 -5.086418e+18
#>   [91] -8.980040e+18  3.741411e+18 -8.786562e+18 -5.803002e+18 -3.023921e+16
#>   [96] -1.924666e+18  4.736079e+18 -1.909284e+18  4.321441e+18 -4.192878e+18
#>  [101] -6.449797e+17  8.035660e+18  6.344654e+18 -4.923027e+17  3.871282e+18
#>  [106] -2.330686e+18  1.719190e+18 -5.118539e+18 -3.641339e+18 -7.059852e+18
#>  [111]  3.012981e+18 -5.415593e+18 -9.187098e+18 -6.167060e+18  8.307970e+18
#>  [116] -4.973544e+18 -4.878023e+18  8.733431e+18  2.023639e+18 -7.485996e+18
#>  [121]  1.255879e+18  6.354789e+18  7.810333e+18  8.586305e+18  7.893625e+18
#>  [126] -6.361564e+18 -8.406181e+18 -3.610570e+18 -2.294288e+18 -6.020644e+18
#>  [131]  3.138434e+18  4.738437e+18  7.372291e+18  8.003593e+18  4.354357e+16
#>  [136]  5.795265e+18  6.502497e+18 -6.716227e+17  3.702705e+18  6.527142e+18
#>  [141] -2.107181e+18  1.344977e+18  8.863477e+18 -9.092288e+18  3.408885e+18
#>  [146] -4.157054e+18 -3.175521e+18  2.061352e+18 -5.742299e+18  8.509479e+18
#>  [151]  6.876026e+18  6.669559e+18 -4.320649e+18  1.778906e+17 -7.783891e+18
#>  [156]  3.779617e+18  4.502898e+17 -9.097948e+18  7.637909e+18  5.976443e+18
#>  [161] -3.062868e+18  1.041203e+18 -4.440008e+18 -2.204435e+18 -3.949455e+18
#>  [166]  7.290393e+18  2.704821e+18 -2.505756e+18  4.317485e+18  5.406432e+18
#>  [171] -2.708934e+18  8.454507e+18  3.789653e+18  8.841724e+18 -8.730132e+18
#>  [176] -2.403490e+18  9.019417e+18  6.029424e+18  3.408256e+18  7.625578e+18
#>  [181] -2.638775e+18 -5.451093e+18 -2.623840e+17 -8.209144e+18 -2.451695e+18
#>  [186] -5.429418e+18 -9.023774e+18 -3.838803e+18 -4.516706e+18  4.407963e+18
#>  [191] -6.860051e+18  9.191858e+17  6.503771e+18 -6.898767e+18  3.352147e+18
#>  [196] -4.914676e+18 -3.339674e+18 -8.100171e+18  8.901147e+18 -6.434208e+18
#>  [201]  7.579780e+18 -8.978641e+17 -3.421074e+18  7.878488e+18 -4.460155e+18
#>  [206] -1.919890e+18 -4.459105e+18  1.665793e+17 -2.935461e+18 -3.861059e+18
#>  [211] -1.074523e+18  8.652912e+18  4.075601e+18 -4.074586e+18  8.648951e+18
#>  [216] -3.959060e+17  2.707237e+18  4.742220e+18 -6.402886e+18 -3.887879e+18
#>  [221]  1.257256e+18  7.896923e+18  1.046374e+18  1.346407e+18 -8.162370e+18
#>  [226] -1.227250e+18  6.088669e+18 -4.342520e+18  9.002610e+18  7.269212e+18
#>  [231]  5.651672e+18 -5.320848e+18  4.237340e+18  1.295590e+18  2.795190e+18
#>  [236] -5.973269e+18  4.517879e+18 -8.174566e+18 -6.423372e+18  8.031892e+18
#>  [241]  7.802041e+17 -7.836044e+18 -4.683360e+18  7.158556e+18  1.311839e+18
#>  [246] -7.058015e+18  2.527205e+17  1.011800e+17  4.129286e+18 -3.697184e+18
#>  [251] -1.296914e+18 -8.470206e+18 -3.381035e+18 -3.028679e+18  4.042806e+18
#>  [256]  9.798029e+17  6.320359e+17  3.923670e+18 -3.766530e+18  4.104047e+17
#>  [261]  6.093894e+18 -6.988230e+18  8.164283e+18 -1.420712e+18  2.647355e+18
#>  [266] -4.086090e+18  5.004175e+18 -4.131988e+18  3.160279e+17 -8.605234e+17
#>  [271]  5.455819e+18 -7.753142e+17  1.830852e+18 -3.590485e+18  4.362766e+18
#>  [276] -6.825155e+18 -6.321970e+17  6.162861e+17 -4.617196e+18  5.568103e+18
#>  [281] -1.224479e+18  1.961338e+18 -4.945917e+18 -2.346535e+18  9.078742e+18
#>  [286] -5.660410e+18  6.609113e+18  3.528621e+18 -2.907106e+17 -8.774087e+18
#>  [291] -7.915943e+18 -4.293875e+18  2.076666e+18  8.885327e+18 -6.986499e+17
#>  [296] -3.442316e+18 -2.434510e+18  2.226226e+16 -7.588264e+18 -7.254113e+18
#>  [301]  8.049527e+17 -3.526232e+18  2.094795e+18  2.595082e+18 -4.825402e+18
#>  [306] -5.497215e+18  5.067903e+18  2.743195e+18  7.107398e+18  5.965920e+18
#>  [311] -1.917279e+18  5.205272e+18  9.094116e+18  3.213004e+18 -2.791508e+18
#>  [316]  2.184645e+18  2.940118e+18  3.013987e+18 -4.165295e+18  7.629440e+18
#>  [321] -3.848151e+18 -4.525428e+18  3.474430e+18 -2.175734e+18  7.588335e+18
#>  [326]  5.961845e+18 -8.340197e+18 -2.228564e+18 -6.042997e+18 -6.273798e+18
#>  [331]  1.465648e+18  8.805307e+18 -2.708870e+17 -5.550265e+18 -2.772519e+18
#>  [336]  7.726727e+18 -2.317538e+18 -7.569079e+18 -3.709486e+18 -4.527268e+18
#>  [341]  5.880838e+18 -5.218442e+18 -3.651382e+18 -8.942324e+18  3.407063e+18
#>  [346]  3.999896e+18  8.623283e+17  8.345790e+18  8.489048e+18 -4.522640e+16
#>  [351] -2.025780e+18 -4.220063e+18  8.526186e+18  5.403263e+18 -3.472726e+18
#>  [356] -8.705440e+18  8.410741e+18 -3.201428e+18 -8.649703e+18 -5.449030e+18
#>  [361] -3.307842e+18  2.290638e+18  3.272023e+17 -8.691941e+17  1.259501e+18
#>  [366]  7.956345e+18  2.815101e+18  2.110830e+18  6.111691e+18  9.086337e+18
#>  [371]  3.677601e+18 -3.200977e+18  5.623766e+18  8.102361e+18  1.762070e+18
#>  [376] -2.159443e+18 -3.892195e+18 -3.360058e+18  1.099953e+18 -2.795334e+18
#>  [381] -5.942744e+18 -8.856908e+18 -2.698319e+18  2.499837e+18 -2.732667e+18
#>  [386]  4.663116e+18 -5.864656e+18 -2.193814e+18  1.460253e+18  4.858969e+18
#>  [391]  9.174145e+18 -1.581465e+18  3.247186e+18 -7.296236e+18 -1.575760e+18
#>  [396]  7.659350e+17 -2.902421e+18 -8.449582e+18 -4.005003e+18 -4.930946e+18
#>  [401]  4.619558e+18 -8.237002e+18  1.052594e+18 -5.325713e+18 -4.623573e+18
#>  [406]  3.031893e+18 -7.465047e+18  2.872040e+18 -6.581500e+18 -4.622593e+18
#>  [411] -5.572463e+18  8.225359e+18 -2.742266e+18  5.021310e+18 -8.062713e+18
#>  [416]  5.591783e+18 -1.813715e+18 -6.088509e+17  4.886487e+18 -8.366700e+18
#>  [421]  4.951797e+18  6.890279e+18 -8.917137e+18  3.113355e+18  2.704532e+18
#>  [426]  5.702457e+18  7.767903e+18 -5.100539e+18 -2.086141e+18  3.393317e+18
#>  [431] -8.994356e+18 -1.568548e+18  3.149053e+18 -1.205262e+18 -6.867950e+18
#>  [436] -5.859876e+18 -1.083828e+18  8.626175e+18  3.220777e+18  8.680948e+18
#>  [441]  2.552408e+17 -5.283802e+18  8.633512e+18 -3.497932e+18  7.716558e+18
#>  [446] -4.399983e+18  3.550457e+18  1.068834e+18  2.663000e+18 -7.982799e+18
#>  [451]  6.135751e+18 -6.277787e+18  8.766198e+18  1.032734e+18 -7.644119e+18
#>  [456]  1.207977e+18  8.629382e+18 -2.098054e+18  5.622386e+18 -8.803248e+17
#>  [461] -7.207825e+18 -7.800967e+18 -6.525509e+18  8.829518e+18 -7.845875e+18
#>  [466]  3.056438e+18 -7.334310e+18 -3.137887e+18 -8.058419e+17  2.672136e+18
#>  [471] -4.812371e+18  8.072276e+18  5.644991e+18 -3.183025e+18  7.118979e+18
#>  [476]  5.603951e+18 -1.654931e+18  1.908270e+18 -3.933360e+17  4.078032e+16
#>  [481]  3.012160e+18 -8.773373e+18 -2.565963e+18 -9.319287e+17  8.330444e+18
#>  [486]  8.798496e+18 -4.366857e+18 -6.195432e+18  4.016076e+18  2.414884e+18
#>  [491] -4.179865e+18  7.933871e+18  1.666141e+18  8.765314e+18 -9.165339e+17
#>  [496]  1.838375e+18  4.967024e+17 -5.073311e+18 -7.426660e+18 -2.432942e+18
#>  [501] -5.578031e+18 -7.328740e+17  3.060357e+18 -3.389859e+18  2.641870e+18
#>  [506] -4.376058e+18 -1.085978e+18 -8.726912e+17  2.213171e+18 -4.196048e+18
#>  [511]  3.232141e+18  8.099402e+18 -7.969422e+18 -2.384991e+17  5.955419e+18
#>  [516]  7.648902e+18 -2.991290e+18  1.795801e+18 -3.513966e+18  7.812415e+18
#>  [521] -4.591547e+18  5.534577e+18  3.673835e+18 -4.449870e+18  1.838057e+18
#>  [526]  1.568795e+18 -5.187323e+18 -5.118997e+18  8.583193e+18  8.211026e+18
#>  [531]  4.667927e+18 -3.904125e+18  1.060983e+18  6.762446e+18 -1.214484e+18
#>  [536]  7.869631e+18 -5.558754e+18 -6.015367e+18  2.455512e+18  3.350153e+18
#>  [541] -8.387773e+17 -3.752283e+18  1.741279e+18 -2.861973e+18  7.639282e+18
#>  [546] -3.432643e+18 -2.926603e+18 -5.197028e+18 -8.360113e+18 -1.728751e+18
#>  [551] -2.073281e+18 -1.872717e+18 -2.351074e+18  6.653423e+18 -7.570833e+18
#>  [556] -8.588300e+18 -3.136795e+18  8.442994e+18 -1.563624e+18  6.041641e+18
#>  [561] -3.803675e+18 -4.707992e+18  3.851136e+18 -1.592816e+18 -9.003902e+18
#>  [566] -8.094355e+18  8.186248e+18  7.723375e+18  5.691146e+18  8.910038e+18
#>  [571] -7.401302e+18  5.193119e+17  6.846097e+18 -8.428460e+18 -1.482938e+18
#>  [576] -9.686395e+17 -5.669608e+18 -5.923953e+18  6.259025e+18 -3.718261e+18
#>  [581]  3.027057e+18 -2.298524e+18 -8.360929e+18  1.307817e+18  1.845232e+17
#>  [586] -5.504878e+18  2.363959e+18  6.623639e+18 -8.053924e+18  6.468887e+18
#>  [591]  3.503668e+18  1.450545e+18 -9.173924e+18  3.527202e+18  4.592221e+18
#>  [596]  2.875599e+18 -3.772658e+17 -7.138659e+18 -7.155669e+18  1.837295e+18
#>  [601] -4.853766e+18  3.770750e+18  4.650694e+18  2.047461e+18 -3.557347e+17
#>  [606] -1.901870e+18  5.404328e+18  7.952783e+18 -1.253441e+18 -3.698366e+18
#>  [611] -2.288751e+18 -5.034532e+18  1.716214e+18 -1.539919e+18 -2.003043e+18
#>  [616] -3.591693e+18  6.668603e+18  5.157762e+18 -5.849138e+18  3.093210e+18
#>  [621]  6.219276e+17 -3.260498e+17 -1.195476e+18  3.689035e+17  3.878503e+18
#>  [626]  4.514457e+18  7.120354e+18 -7.242724e+18 -3.438264e+18 -6.197227e+18
#>  [631]  5.170207e+18  5.140460e+18 -5.983315e+18  2.922450e+18 -8.672127e+18
#>  [636] -1.761235e+17 -8.778135e+17  6.730010e+18  5.662571e+17  4.320434e+18
#>  [641]  1.469754e+18  4.568186e+18 -8.592483e+18 -2.904376e+18  3.385952e+18
#>  [646] -7.500248e+18 -6.092003e+18 -4.747082e+16 -6.415640e+18 -2.103755e+18
#>  [651] -8.351123e+18 -2.839243e+18 -5.390108e+18  7.511624e+18 -4.996151e+18
#>  [656] -4.657431e+18 -5.746908e+18 -2.713804e+18  6.517033e+18  5.730166e+18
#>  [661] -8.583651e+18 -2.730181e+18 -3.238545e+18 -3.347935e+18  1.118661e+18
#>  [666]  5.032870e+18 -1.505446e+18  4.284516e+18  6.697200e+18  2.532199e+18
#>  [671]  2.816382e+18  3.330613e+18 -3.993806e+17  8.125267e+17 -7.764829e+18
#>  [676]  2.118928e+18  1.213316e+18  5.393073e+18 -2.672268e+18 -7.327581e+17
#>  [681]  6.035023e+18  8.243953e+18  1.399231e+18 -8.229932e+18  7.748784e+18
#>  [686] -7.628243e+18 -3.009405e+18  2.760705e+18 -6.171255e+18 -2.140938e+18
#>  [691]  1.226380e+18  9.167641e+18 -4.116252e+18  7.933511e+18  8.823855e+18
#>  [696]  7.846540e+18 -8.620721e+18  7.206225e+18  7.624183e+18  2.147097e+18
#>  [701]  8.637823e+18 -6.966665e+18  4.639143e+18  4.808640e+18 -7.531385e+18
#>  [706] -7.772136e+18 -1.439588e+18  8.467175e+18 -6.432000e+18  5.832186e+18
#>  [711]  3.232249e+18  7.634966e+18 -2.737381e+18  1.420633e+18 -4.305729e+18
#>  [716] -8.808059e+18  8.020759e+18  7.721935e+17  2.216080e+18  2.975845e+18
#>  [721] -2.618051e+18 -2.850214e+18 -7.227802e+18 -6.220266e+18 -8.728342e+18
#>  [726] -3.623945e+18 -6.403193e+18  1.559110e+18  5.151637e+18 -1.062445e+18
#>  [731]  2.696241e+18  8.601634e+18  2.952225e+18  7.858719e+18  2.631910e+17
#>  [736] -1.157116e+18  1.675740e+18  1.905465e+18  2.840223e+18 -7.313954e+18
#>  [741]  3.271523e+18 -1.019529e+18  7.063997e+17 -1.253086e+18  3.029378e+18
#>  [746] -4.827318e+18 -6.773547e+17  5.561256e+18  5.686451e+18  6.347412e+18
#>  [751] -6.240077e+18 -4.763945e+18  8.819211e+18 -8.327569e+18  6.486470e+18
#>  [756]  7.704969e+18  3.464748e+18 -7.699905e+18  4.557103e+18  8.520556e+18
#>  [761]  7.475079e+18 -3.526567e+18 -3.805003e+18 -8.245602e+18 -5.915067e+17
#>  [766]  3.883098e+18 -3.115953e+18  3.988478e+18 -6.146835e+18  4.490006e+18
#>  [771] -3.670188e+18  3.296099e+18 -2.885217e+18  1.583821e+18 -6.773034e+16
#>  [776]  2.946261e+18  4.518318e+18 -7.791375e+18 -8.149638e+18  7.811093e+18
#>  [781]  7.716692e+18  9.168315e+18  2.855332e+18 -6.162214e+18 -1.608080e+18
#>  [786]  2.608799e+18  7.320764e+18 -5.369098e+17 -4.921649e+18 -2.359826e+18
#>  [791] -7.657881e+18 -7.859294e+17  8.943235e+18  7.078328e+18  2.634279e+17
#>  [796] -4.305353e+18 -1.276519e+18 -3.380279e+18 -3.670818e+18 -8.116937e+18
#>  [801]  2.690973e+18  4.552155e+18  3.496960e+18 -2.076441e+18  6.667740e+18
#>  [806]  6.336777e+18 -1.430642e+18  5.265647e+18 -8.948623e+18  6.596897e+18
#>  [811]  8.528919e+18 -3.085465e+18 -1.061287e+18 -4.535313e+17  7.731812e+18
#>  [816] -3.418716e+18  1.836065e+18 -5.034474e+18 -5.104640e+18 -5.059849e+18
#>  [821]  2.829624e+18  5.478075e+18 -3.481306e+18  5.524403e+18  9.085541e+18
#>  [826]  7.104796e+18 -4.701454e+18 -1.206337e+18  4.999523e+18  1.917262e+18
#>  [831]  8.412738e+18  5.034023e+17 -4.268850e+18 -5.854994e+18  7.011881e+18
#>  [836] -2.634804e+18 -5.477636e+18  2.704924e+18  2.686690e+18 -4.748863e+18
#>  [841]  4.041523e+18 -5.099493e+18 -2.448884e+18  8.615456e+18  1.936178e+18
#>  [846] -1.951970e+16  7.337515e+18  8.323022e+18  8.815262e+18 -3.104129e+18
#>  [851] -7.103960e+18 -2.634808e+18 -6.787290e+18 -8.182976e+18  3.805375e+18
#>  [856] -4.252303e+18  4.648001e+18 -7.124328e+18  1.233295e+18 -3.123955e+18
#>  [861]  7.038976e+18  6.716804e+18  8.929062e+18  2.560283e+18 -2.799346e+18
#>  [866]  3.805325e+17  5.729276e+18 -5.568396e+17  8.679506e+18  8.846672e+18
#>  [871] -1.341094e+18  1.036358e+18  6.841972e+18  7.332658e+18 -3.826235e+18
#>  [876]  6.095193e+18  6.427347e+18  5.652173e+18 -7.742037e+18  6.612858e+18
#>  [881]  2.529734e+18 -7.590607e+18  4.926817e+18  3.248271e+18 -6.182075e+18
#>  [886]  6.233129e+18  2.798772e+18 -4.149099e+18  8.519886e+18 -2.558308e+18
#>  [891] -8.410280e+18  7.648822e+18 -4.632058e+18 -4.083520e+18 -8.155477e+18
#>  [896]  8.356472e+17 -5.371290e+18  6.843695e+18 -5.007037e+18 -1.097176e+17
#>  [901] -7.790372e+18 -4.661985e+18  2.701394e+18 -9.076522e+17 -3.141306e+18
#>  [906] -7.611509e+18  3.507200e+16  9.154389e+17 -9.199047e+18  6.298278e+18
#>  [911]  7.406805e+16 -2.356935e+18  5.254567e+17 -1.556672e+18 -1.114366e+17
#>  [916]  2.276496e+18  6.256437e+18 -6.548841e+18 -8.938182e+18  2.318810e+18
#>  [921]  3.764700e+18 -1.770896e+18 -6.080515e+18  4.526855e+18 -5.643969e+18
#>  [926]  3.200987e+18  1.634801e+18  4.592278e+18 -8.590064e+18  8.089392e+18
#>  [931] -7.455803e+18 -6.233942e+18 -3.846861e+17  1.204302e+18  7.699075e+18
#>  [936]  9.887566e+16 -9.051709e+18  7.024319e+18 -3.950567e+18 -9.051237e+18
#>  [941]  5.483883e+18 -5.818266e+18  6.972934e+18 -8.669544e+18  5.253824e+18
#>  [946]  2.351982e+18 -3.006609e+18  4.464864e+17 -6.256863e+18 -7.944853e+18
#>  [951]  1.829107e+18  7.524632e+18  6.284408e+18 -1.629171e+18  3.272813e+18
#>  [956] -3.610647e+18 -5.707746e+18  9.067542e+18 -1.809073e+18 -6.715030e+18
#>  [961] -4.752720e+18 -4.442432e+18 -7.906244e+17 -4.218349e+18 -5.873846e+18
#>  [966] -8.617336e+18  4.726808e+17  9.118786e+18 -1.908160e+18 -5.552392e+18
#>  [971] -3.599210e+18 -7.063610e+18 -3.021285e+18 -5.606306e+18 -6.139783e+18
#>  [976] -5.676607e+18 -7.409496e+18 -2.906137e+17  5.925825e+18 -2.796650e+18
#>  [981] -9.168781e+18 -9.756081e+17  6.553911e+18 -1.055293e+18 -3.037836e+18
#>  [986] -5.449944e+17 -6.367237e+18 -5.925370e+18 -3.363091e+18  5.269696e+18
#>  [991] -7.052181e+18 -4.327140e+18 -1.103039e+18  3.807848e+18 -8.042194e+17
#>  [996]  4.193495e+18  1.897810e+18 -1.866616e+18 -7.706012e+18  3.790470e+18
xorshift_128(123456789, 362436069, 521288629, 88675123, n = 1000)
#>    [1] 1.820123e+11 9.023208e+11 5.432700e+11 2.947180e+11 3.718922e+14
#>    [6] 2.186712e+15 1.183813e+15 1.708913e+15 7.645033e+17 3.784016e+18
#>   [11] 1.514136e+18 2.712351e+18 3.508093e+18 1.743529e+18 1.687072e+18
#>   [16] 1.399182e+18 6.517774e+18 3.447899e+18 8.472362e+18 3.575487e+18
#>   [21] 7.515537e+17 1.252473e+18 2.959020e+17 4.554714e+18 4.994853e+18
#>   [26] 6.432825e+18 8.685980e+18 1.506914e+18 2.608979e+18 5.969540e+18
#>   [31] 8.942577e+18 2.712616e+18 6.464631e+18 3.781890e+18 7.249937e+18
#>   [36] 7.495607e+18 8.756371e+18 7.913560e+18 1.249098e+18 6.116724e+18
#>   [41] 7.351744e+17 9.358657e+17 5.734981e+18 9.338844e+17 7.474758e+18
#>   [46] 8.235444e+18 8.633591e+18 3.031056e+18 8.032226e+18 5.509853e+18
#>   [51] 4.698313e+18 1.679018e+18 5.045050e+18 4.819122e+18 7.069480e+18
#>   [56] 2.136474e+18 5.119613e+18 1.512523e+17 4.686390e+18 7.979576e+18
#>   [61] 5.600277e+18 8.840766e+18 8.108373e+18 6.165091e+17 8.920292e+18
#>   [66] 8.876560e+18 4.484785e+18 4.969666e+18 7.416708e+18 7.116718e+18
#>   [71] 5.361862e+18 5.692789e+18 4.783829e+18 4.437861e+18 4.517182e+18
#>   [76] 8.589070e+18 2.996617e+18 6.347055e+18 1.950520e+18 5.616056e+17
#>   [81] 7.005740e+18 7.853687e+18 3.894000e+17 6.573059e+18 1.060257e+18
#>   [86] 8.282345e+18 5.531513e+18 3.023179e+18 7.977515e+18 7.337940e+18
#>   [91] 3.915081e+18 4.885394e+18 8.945124e+18 5.400021e+18 3.058136e+18
#>   [96] 7.663975e+17 7.876807e+18 6.755209e+18 9.431544e+17 1.295703e+18
#>  [101] 2.214829e+17 6.600415e+18 2.250447e+18 3.106559e+18 4.715341e+18
#>  [106] 3.343929e+18 1.664959e+18 2.770654e+18 1.932067e+18 8.423805e+18
#>  [111] 4.956987e+18 5.549362e+17 7.087360e+18 3.179514e+18 4.526067e+18
#>  [116] 6.546971e+18 2.152169e+18 3.542639e+18 8.157847e+18 6.097576e+17
#>  [121] 1.878479e+18 7.249854e+18 6.457251e+18 2.020438e+18 8.364422e+18
#>  [126] 7.348505e+17 5.336683e+18 4.336110e+17 3.380297e+18 5.688261e+18
#>  [131] 8.746289e+18 6.680990e+18 4.321004e+18 9.250146e+17 9.060167e+18
#>  [136] 7.191349e+18 2.158452e+18 6.692928e+18 4.588721e+18 4.183281e+18
#>  [141] 8.836789e+18 3.926485e+18 8.850881e+18 3.418692e+18 4.689803e+18
#>  [146] 7.340480e+18 5.043768e+18 1.746468e+18 7.886937e+17 7.223861e+18
#>  [151] 2.774389e+18 2.655120e+18 4.734894e+18 2.875765e+18 4.092727e+17
#>  [156] 1.844953e+18 7.236227e+17 7.756728e+18 2.180890e+18 3.218631e+18
#>  [161] 8.100688e+18 3.529667e+18 1.016887e+18 8.489262e+18 6.839789e+18
#>  [166] 5.681201e+18 6.524082e+18 5.771531e+18 5.804618e+18 6.690757e+18
#>  [171] 6.325907e+18 4.360607e+18 2.091633e+18 7.823654e+18 7.659096e+18
#>  [176] 5.263944e+18 7.198748e+18 7.574618e+18 6.337154e+18 8.400440e+18
#>  [181] 3.431723e+18 5.368640e+18 8.217478e+18 6.604574e+18 8.400773e+18
#>  [186] 4.016272e+18 1.684127e+18 9.831759e+17 3.078835e+18 5.481862e+17
#>  [191] 1.651007e+18 4.390282e+18 4.040542e+18 2.031617e+18 5.110905e+18
#>  [196] 1.226164e+18 4.729479e+18 3.172941e+18 4.614423e+17 3.962411e+18
#>  [201] 7.264928e+18 9.246322e+17 3.491831e+18 1.380800e+18 1.846925e+18
#>  [206] 5.583987e+18 2.972735e+18 8.540606e+18 7.201875e+18 2.541324e+18
#>  [211] 8.637031e+15 4.935566e+18 5.324796e+18 5.651472e+18 4.971800e+18
#>  [216] 7.222743e+17 7.516539e+18 6.232365e+18 1.599132e+18 3.190819e+18
#>  [221] 4.256231e+18 8.981112e+18 2.241157e+18 8.337554e+18 4.488946e+18
#>  [226] 6.341774e+18 7.590787e+18 4.808654e+18 2.528022e+18 8.022669e+18
#>  [231] 5.113226e+18 2.736967e+18 6.040375e+18 8.260209e+18 7.286788e+18
#>  [236] 7.085740e+18 6.020954e+18 3.462206e+18 6.085424e+18 7.291386e+18
#>  [241] 4.852111e+18 1.353058e+18 2.835071e+18 4.334137e+18 3.587497e+18
#>  [246] 1.950426e+18 1.538733e+17 1.263083e+18 7.709965e+18 2.935560e+17
#>  [251] 1.400785e+18 4.045915e+18 6.280690e+18 4.320626e+18 6.028029e+18
#>  [256] 4.919412e+18 6.378239e+18 3.707813e+18 6.799229e+18 3.489089e+18
#>  [261] 5.251838e+18 2.476109e+18 6.786951e+18 3.735392e+18 7.591634e+18
#>  [266] 6.051911e+18 8.263709e+18 6.319112e+17 5.266893e+18 7.016522e+16
#>  [271] 4.983416e+17 3.026093e+18 2.378621e+18 1.442892e+18 4.684104e+18
#>  [276] 7.034437e+18 6.111569e+18 8.178815e+18 4.248102e+18 6.794546e+18
#>  [281] 8.103415e+18 4.611167e+17 7.014813e+18 7.460942e+18 4.669197e+18
#>  [286] 8.584995e+18 2.677791e+18 1.596659e+18 3.768381e+18 7.135015e+18
#>  [291] 9.470550e+17 6.386082e+18 3.704618e+18 8.465404e+18 6.673719e+18
#>  [296] 3.438135e+17 8.953734e+18 3.447169e+18 7.099756e+18 5.596059e+18
#>  [301] 2.443961e+18 5.172432e+18 2.172817e+18 1.875151e+18 7.931649e+18
#>  [306] 7.531747e+18 5.597277e+18 8.691209e+18 9.112318e+18 2.849094e+18
#>  [311] 4.449637e+17 7.647988e+18 4.623415e+18 4.021669e+18 6.278770e+18
#>  [316] 2.706924e+18 2.948490e+18 7.049738e+18 2.587894e+18 9.132541e+18
#>  [321] 1.083051e+18 4.361250e+18 5.711067e+18 2.594815e+18 1.521944e+18
#>  [326] 1.992481e+18 6.552165e+18 7.767868e+18 8.719366e+18 6.286330e+18
#>  [331] 7.087798e+18 7.028638e+18 1.420175e+18 6.301490e+18 3.311187e+18
#>  [336] 1.849721e+18 6.803476e+18 7.911597e+18 2.454939e+18 7.476573e+18
#>  [341] 8.055815e+18 6.858228e+18 1.081126e+18 8.668064e+18 8.611386e+18
#>  [346] 4.758172e+18 5.364792e+18 7.725410e+18 1.333882e+18 1.434979e+18
#>  [351] 4.242001e+18 2.263513e+18 1.934609e+18 6.483880e+18 7.494659e+18
#>  [356] 4.310477e+18 1.602578e+18 7.688608e+18 1.283469e+18 6.536085e+18
#>  [361] 6.878822e+18 5.860421e+18 4.510049e+18 4.425593e+18 3.319352e+18
#>  [366] 2.564871e+18 6.182571e+18 4.478175e+18 7.654170e+18 8.378558e+18
#>  [371] 5.080605e+18 6.184763e+18 6.363540e+17 5.197376e+18 1.138040e+17
#>  [376] 1.024455e+18 6.848016e+18 1.236992e+18 5.592575e+18 6.300635e+16
#>  [381] 1.698271e+18 6.021058e+18 5.180254e+17 4.814272e+17 7.304984e+18
#>  [386] 5.622466e+18 3.861975e+18 4.736036e+18 2.349933e+18 6.265797e+18
#>  [391] 6.878482e+18 3.420537e+18 1.531182e+18 1.774070e+18 1.308680e+18
#>  [396] 2.335548e+17 1.667820e+18 4.540492e+17 6.832329e+18 5.478356e+18
#>  [401] 8.166026e+18 2.281856e+18 3.807630e+18 4.655001e+18 5.930090e+18
#>  [406] 2.011690e+18 7.775573e+18 2.011395e+18 1.649714e+18 6.552380e+18
#>  [411] 8.231530e+18 2.794237e+18 1.710610e+18 4.081596e+18 6.281505e+18
#>  [416] 5.256880e+18 5.387216e+18 6.310740e+18 7.174498e+18 8.435215e+18
#>  [421] 2.772968e+18 3.178371e+18 4.030462e+18 4.454227e+18 4.047350e+18
#>  [426] 3.854145e+18 8.811100e+18 4.365568e+18 6.648802e+18 8.325495e+18
#>  [431] 3.676263e+18 6.693715e+18 3.064683e+18 7.136429e+17 2.260580e+18
#>  [436] 7.282784e+18 8.072831e+18 4.841117e+18 6.493142e+18 5.606035e+18
#>  [441] 9.105972e+18 5.016537e+18 1.469241e+17 3.097020e+18 6.656647e+18
#>  [446] 1.497318e+18 6.437554e+18 6.542424e+18 1.132930e+18 2.810986e+18
#>  [451] 4.031493e+18 4.028450e+18 1.975834e+16 3.638617e+18 8.054401e+18
#>  [456] 7.441630e+18 6.242767e+18 7.830671e+18 3.811607e+18 8.952457e+18
#>  [461] 4.379059e+18 3.563572e+18 3.019691e+18 5.071488e+18 6.264412e+18
#>  [466] 4.234798e+18 6.020596e+18 1.872453e+18 3.635122e+18 2.348081e+18
#>  [471] 1.766443e+18 2.077613e+18 5.000558e+18 3.033305e+18 3.443540e+18
#>  [476] 7.348455e+18 9.053263e+17 1.933839e+18 8.804862e+18 4.024556e+18
#>  [481] 5.055326e+18 1.177170e+18 2.066611e+18 3.122764e+17 2.499579e+17
#>  [486] 6.754943e+18 3.594884e+18 7.023320e+18 6.771121e+18 9.315645e+17
#>  [491] 2.422689e+18 1.412090e+17 2.104606e+18 9.011973e+18 6.602853e+18
#>  [496] 6.208252e+17 4.783135e+18 5.125428e+18 9.122336e+17 1.673367e+18
#>  [501] 6.769966e+18 1.437176e+18 6.354474e+18 8.588816e+18 5.259291e+18
#>  [506] 3.057228e+18 1.108828e+18 8.275830e+17 6.496904e+18 1.990857e+18
#>  [511] 1.017350e+18 2.007145e+18 9.350268e+17 1.209342e+18 1.439922e+18
#>  [516] 2.418990e+18 2.136106e+18 5.668813e+18 9.193827e+18 3.350654e+18
#>  [521] 1.452942e+18 5.611468e+17 4.598404e+18 1.238551e+18 5.357409e+18
#>  [526] 7.992615e+16 5.184674e+18 2.949850e+18 6.305277e+18 8.567727e+18
#>  [531] 6.017656e+18 3.563282e+17 6.169500e+18 1.605587e+18 6.742343e+18
#>  [536] 2.999085e+18 8.078376e+18 2.846462e+18 6.603508e+17 2.914318e+18
#>  [541] 5.098772e+18 7.219310e+18 4.399881e+18 7.438925e+18 3.631415e+18
#>  [546] 2.931350e+18 7.620161e+18 1.406240e+18 7.900816e+17 6.050812e+18
#>  [551] 4.058130e+18 3.774489e+18 8.526067e+18 2.041190e+18 5.858979e+18
#>  [556] 7.528851e+18 8.404566e+18 6.782033e+17 2.467558e+18 7.719449e+18
#>  [561] 5.625376e+17 4.996546e+18 7.819772e+18 1.620386e+16 8.341518e+18
#>  [566] 8.301994e+18 3.799076e+18 5.603420e+17 7.808201e+18 6.085567e+18
#>  [571] 6.271868e+18 7.286264e+18 1.486746e+18 2.096128e+18 1.958961e+18
#>  [576] 8.192189e+18 7.697096e+18 4.604496e+18 6.460565e+18 6.047742e+18
#>  [581] 5.366236e+18 5.059661e+18 6.613647e+18 7.436122e+18 1.680949e+18
#>  [586] 1.345011e+18 7.472670e+17 1.310450e+17 8.537337e+18 3.969024e+18
#>  [591] 3.759832e+18 5.112900e+18 1.916507e+18 8.390077e+18 3.228292e+18
#>  [596] 3.496960e+18 1.415570e+18 1.986338e+18 6.812432e+18 6.021816e+18
#>  [601] 7.558335e+18 8.615987e+17 1.118622e+17 2.562536e+18 8.064259e+18
#>  [606] 3.722018e+18 6.417779e+18 3.885239e+17 4.230055e+18 3.676693e+18
#>  [611] 1.293855e+18 3.905934e+18 5.942636e+18 5.809412e+18 1.968060e+18
#>  [616] 8.627031e+18 1.851099e+18 5.758613e+18 3.126177e+18 7.610323e+18
#>  [621] 9.274374e+17 1.625673e+18 3.358530e+18 5.971776e+18 6.241504e+18
#>  [626] 4.332428e+18 3.629741e+18 6.989279e+18 4.322789e+18 5.140209e+17
#>  [631] 3.592746e+18 6.452629e+18 8.118997e+18 7.404580e+18 8.577727e+18
#>  [636] 5.477134e+18 6.881251e+18 3.106826e+18 9.795763e+17 6.071629e+18
#>  [641] 8.984585e+17 3.883984e+18 4.710325e+17 4.942918e+18 6.414564e+17
#>  [646] 6.108948e+17 4.986148e+18 4.036701e+18 5.465943e+17 2.493204e+18
#>  [651] 7.670011e+17 1.964765e+18 5.502773e+18 6.644539e+18 8.197867e+18
#>  [656] 5.423362e+18 1.624286e+18 3.226285e+17 5.804006e+18 4.800737e+17
#>  [661] 4.943318e+18 7.967610e+18 7.052138e+18 3.151963e+18 4.700734e+18
#>  [666] 4.993724e+18 3.085523e+18 1.052156e+18 5.972589e+18 9.056676e+18
#>  [671] 2.812616e+18 4.589288e+17 4.888288e+18 4.632363e+18 2.627871e+18
#>  [676] 3.339735e+18 2.880368e+18 3.175729e+18 3.971386e+18 3.339842e+18
#>  [681] 4.517724e+18 9.131271e+18 6.911190e+18 4.780235e+18 1.312984e+18
#>  [686] 6.155586e+18 4.703001e+18 5.374475e+18 7.077746e+18 6.855342e+18
#>  [691] 4.418038e+18 2.875192e+18 8.346934e+18 3.763384e+18 8.567266e+18
#>  [696] 7.262624e+18 6.556546e+18 4.678777e+18 2.165269e+18 3.081583e+18
#>  [701] 7.190429e+18 5.891739e+18 3.105054e+18 2.178153e+18 3.573261e+18
#>  [706] 9.029844e+18 1.379260e+18 2.307123e+18 6.356919e+18 6.413481e+18
#>  [711] 7.709460e+18 7.972261e+18 8.300806e+17 6.399588e+18 2.431534e+18
#>  [716] 6.209476e+18 8.486407e+18 5.914518e+18 8.657393e+18 5.607178e+18
#>  [721] 1.620993e+18 2.245466e+18 5.673029e+18 8.732759e+18 7.478457e+18
#>  [726] 3.695259e+18 6.293812e+18 6.225792e+18 8.577068e+18 4.086399e+17
#>  [731] 7.891273e+18 5.863605e+17 4.719946e+18 1.895578e+18 7.837144e+18
#>  [736] 9.055784e+18 4.072909e+18 5.851167e+18 2.633524e+18 4.460690e+18
#>  [741] 3.060576e+18 1.777156e+18 6.710830e+18 6.680307e+18 4.892427e+18
#>  [746] 1.645080e+18 5.108101e+18 5.484309e+18 2.644300e+18 7.938027e+18
#>  [751] 3.776226e+18 7.397123e+18 3.325383e+18 8.917384e+17 4.880045e+17
#>  [756] 6.837709e+18 4.743414e+18 5.550019e+18 7.351940e+18 2.016887e+18
#>  [761] 4.187627e+18 6.521158e+18 4.729532e+17 7.201625e+17 2.860853e+18
#>  [766] 8.943366e+18 5.164884e+17 3.664891e+17 4.849833e+18 2.956424e+18
#>  [771] 8.720710e+18 3.718823e+18 3.126631e+16 1.433822e+18 6.616197e+18
#>  [776] 5.253735e+18 3.484242e+18 8.770697e+17 2.561046e+18 2.540133e+18
#>  [781] 4.097896e+18 7.743750e+18 2.143868e+18 4.344493e+18 1.039346e+18
#>  [786] 2.328378e+18 4.173239e+18 5.775235e+18 4.458510e+18 7.084771e+18
#>  [791] 6.696641e+17 8.575057e+18 5.433397e+18 5.155930e+18 1.614733e+18
#>  [796] 7.265176e+18 1.562312e+18 4.490618e+18 7.862552e+18 7.963984e+18
#>  [801] 6.386575e+17 5.059211e+18 4.578254e+18 9.170792e+18 8.055357e+18
#>  [806] 8.754886e+18 1.103614e+18 6.473238e+18 7.287961e+18 2.237090e+18
#>  [811] 7.586983e+18 7.139729e+18 2.754219e+18 7.249314e+18 6.837520e+18
#>  [816] 7.598332e+18 8.588787e+18 4.175450e+18 8.839452e+18 8.768824e+18
#>  [821] 9.041892e+18 3.047674e+18 3.508036e+18 4.473665e+18 7.407716e+18
#>  [826] 1.219013e+18 6.371523e+18 3.809034e+18 4.172844e+18 9.163503e+18
#>  [831] 5.111319e+18 7.971799e+18 1.182371e+18 3.821981e+18 8.434398e+17
#>  [836] 7.588594e+18 4.437096e+18 6.612967e+18 8.754525e+17 1.895298e+18
#>  [841] 5.019037e+18 3.585450e+18 8.654562e+17 9.052317e+18 1.359160e+17
#>  [846] 2.350166e+18 4.123204e+18 5.095516e+18 5.887357e+18 7.302889e+18
#>  [851] 7.485600e+18 7.632640e+18 7.451662e+18 3.820257e+18 4.835978e+18
#>  [856] 5.644866e+18 7.345061e+18 8.269515e+18 2.938011e+18 3.302078e+18
#>  [861] 4.548397e+18 6.055758e+18 6.017226e+18 1.969843e+18 2.500627e+18
#>  [866] 2.613104e+18 8.953345e+18 3.094375e+18 6.328469e+18 8.019554e+18
#>  [871] 1.597788e+18 5.192869e+18 8.783469e+18 5.761869e+18 4.201545e+18
#>  [876] 8.399508e+17 6.569422e+18 6.416716e+18 1.448744e+18 6.858718e+18
#>  [881] 6.787052e+18 6.992142e+18 6.708197e+18 8.820971e+18 6.351803e+18
#>  [886] 8.185674e+18 1.257949e+18 4.051496e+18 5.866527e+18 1.554906e+18
#>  [891] 5.932459e+18 6.626075e+18 6.496529e+18 1.232972e+18 2.172886e+18
#>  [896] 2.183349e+18 4.355399e+17 8.508684e+17 2.948506e+18 5.790920e+18
#>  [901] 9.189575e+17 8.093153e+18 9.780809e+16 5.001179e+18 5.762951e+18
#>  [906] 4.980830e+18 6.935839e+18 1.967318e+18 8.918008e+18 4.238798e+18
#>  [911] 5.953559e+18 2.533906e+18 4.723153e+18 2.096908e+18 5.367822e+18
#>  [916] 4.095637e+18 1.832926e+18 3.896106e+18 8.176846e+18 2.253996e+17
#>  [921] 7.263472e+18 2.348636e+18 7.039060e+18 7.256491e+18 7.511245e+18
#>  [926] 8.576303e+18 7.542389e+18 5.924601e+18 3.453301e+18 8.180392e+18
#>  [931] 5.080674e+18 2.910282e+18 7.165277e+18 2.740965e+18 8.192623e+18
#>  [936] 4.813468e+18 6.861036e+18 4.015547e+18 3.005320e+18 9.822560e+17
#>  [941] 1.670194e+18 1.137208e+18 8.191514e+18 8.172778e+18 8.630934e+17
#>  [946] 4.979501e+18 4.928709e+18 7.598181e+18 5.748308e+18 2.351921e+18
#>  [951] 6.261855e+18 5.846934e+18 3.324294e+18 1.378892e+18 8.144642e+18
#>  [956] 2.560050e+17 4.579526e+18 4.241996e+18 8.590070e+18 1.787303e+18
#>  [961] 5.397302e+18 8.880396e+18 4.880452e+18 3.832715e+18 5.127859e+18
#>  [966] 3.424433e+18 5.002946e+18 8.462614e+17 1.528986e+17 2.145480e+18
#>  [971] 2.944107e+18 2.929757e+18 3.238018e+18 2.300672e+17 6.172378e+17
#>  [976] 7.270830e+18 3.865845e+18 4.898187e+18 3.608285e+18 8.054073e+18
#>  [981] 7.748805e+18 1.812317e+18 5.566187e+18 8.805957e+17 3.302663e+18
#>  [986] 1.430506e+17 5.475560e+18 8.936629e+18 3.623515e+17 1.919975e+18
#>  [991] 5.061746e+18 1.165215e+18 3.438131e+18 2.150218e+18 5.994179e+18
#>  [996] 2.173733e+18 8.905737e+18 2.394168e+18 1.078112e+18 5.018828e+18